Update UnionPay regexes
We discovered that there are some bin ranges that are shared between
Maestro and UnionPay. In particular one bin can be either UnionPay or
Maestro depending on the length of the card number. This attempts to
update the regexes to better align with what we believe UnionPay numbers
to be and exclude Maestro cards from being labeled as UnionPay.